fix(case-law): keep tracked items out of slice walks and bound the listing
Review round on the reconciliation loop.

A slice walk treated every unheld identity as fetchable, including ones the
store was already tracking. A tip slice is re-walked daily, so a parked item
was asked for daily whatever its `nextAttemptAt` said — the widening backoff
served nothing — and a terminal item was pulled straight back into the hunt it
had already left, which is the unbounded loop the retirement exists to end.
Both also spent the per-unit ingest budget ahead of misses nothing was
tracking yet. The walk now reads the store for what it found missing and
leaves tracked identities to the due-retry path. A skipped item is still
neither held nor written, so the ledger keeps reading the slice short.

Settledness weighed two populations against each other: `reported` and
`collected` describe the listing as it is now, while the terminal count
described rows written whenever they were written. A publisher that drops an
identity and lists another in its place left a retired row vouching for a
shortfall it had nothing to do with. A complete walk is authoritative about
what a slice lists, so it now prunes retired rows the listing no longer
contains before writing the ledger. Terminal rows only: a parked row keeps its
slice short either way.

`totalPages` came from the publisher with no ceiling, so one wrong value made
a single work unit an unbounded request sequence that only the hard deadline
ended, by killing the process — after which the replacement began the same
walk. Capped, and refused rather than truncated: a partial listing written as
if it were the whole slice can read as fully collected.

The loop's failure streak was shared across sources, so a healthy source
taking its turn reset it every rotation and a broken one never passed its
first doubled delay. Streaks are now per source.

`sum(...)` is bigint and returns a string, so the status rollup compared a
string to a number and reported string values behind a `number` type. All five
aggregates in that handler now cast to int, including two that predate this
branch. `listingIdentityKey` no longer formats a key from an empty component,
which its own parser rejected, so the round trip holds for every key it emits.

Two suppression directives named a rule that had nothing to report at their
site, which the unused-directive check treats as an error.

+++ apps/api/src/handlers/case-law/ingestion/status.ts
@@ -108,7 +108,7 @@ export const getIngestionStatus = async (
       // oxlint-disable-next-line no-db-await-in-loop/no-db-await-in-loop, no-await-in-loop -- sequential per-source aggregation reads on a single scoped connection
       const [hourRow] = await db
         .select({
-          inserted: sql<number>`coalesce(sum(${caseLawIngestionEvents.inserted}), 0)`,
+          inserted: sql<number>`coalesce(sum(${caseLawIngestionEvents.inserted}), 0)::int`,
         })
         .from(caseLawIngestionEvents)
         .where(
@@ -119,7 +119,7 @@ export const getIngestionStatus = async (
       // oxlint-disable-next-line no-db-await-in-loop/no-db-await-in-loop, no-await-in-loop -- sequential per-source aggregation reads on a single scoped connection
       const [dayRow] = await db
         .select({
-          inserted: sql<number>`coalesce(sum(${caseLawIngestionEvents.inserted}), 0)`,
+          inserted: sql<number>`coalesce(sum(${caseLawIngestionEvents.inserted}), 0)::int`,
         })
         .from(caseLawIngestionEvents)
         .where(
@@ -170,7 +170,7 @@ export const getIngestionStatus = async (
       const [sliceRow] = await db
         .select({
           slices: count(),
-          shortSlices: sql<number>`coalesce(sum(case when ${caseLawCoverageSlices.collected} < ${caseLawCoverageSlices.reported} then 1 else 0 end), 0)`,
+          shortSlices: sql<number>`coalesce(sum(case when ${caseLawCoverageSlices.collected} < ${caseLawCoverageSlices.reported} then 1 else 0 end), 0)::int`,
           lastCheckedAt: sql<Date | null>`max(${caseLawCoverageSlices.checkedAt})`,
         })
         .from(caseLawCoverageSlices)
@@ -179,8 +179,8 @@ export const getIngestionStatus = async (
       // oxlint-disable-next-line no-db-await-in-loop/no-db-await-in-loop, no-await-in-loop -- sequential per-source aggregation reads on a single scoped connection
       const [itemRow] = await db
         .select({
-          parked: sql<number>`coalesce(sum(case when ${caseLawReconciliationItems.status} = ${RECONCILIATION_ITEM_STATUS.PARKED} then 1 else 0 end), 0)`,
-          terminal: sql<number>`coalesce(sum(case when ${caseLawReconciliationItems.status} = ${RECONCILIATION_ITEM_STATUS.TERMINAL} then 1 else 0 end), 0)`,
+          parked: sql<number>`coalesce(sum(case when ${caseLawReconciliationItems.status} = ${RECONCILIATION_ITEM_STATUS.PARKED} then 1 else 0 end), 0)::int`,
+          terminal: sql<number>`coalesce(sum(case when ${caseLawReconciliationItems.status} = ${RECONCILIATION_ITEM_STATUS.TERMINAL} then 1 else 0 end), 0)::int`,
         })
         .from(caseLawReconciliationItems)
         .where(sql`${caseLawReconciliationItems.sourceId} = ${source.id}`);
